agrofinsense/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
motor==3.4.0
redis==5.0.4
celery==5.4.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
easyocr==1.7.1
PyMuPDF==1.24.5
Pillow==10.3.0
xgboost==2.0.3
scikit-learn==1.5.0
statsmodels==0.14.2
torch==2.3.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
geoalchemy2==0.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
